rag-mcp/requirements.txt
mcp[cli]>=1.0.0
networkx>=3.0
python-dotenv>=1.0.0
